moved api responses to common folder
