+++ common/package.json
@@ -49,6 +49,8 @@
     "typescript": "^4.1.3"
   },
   "dependencies": {
+    "class-transformer": "^0.4.0",
+    "class-validator": "^0.13.1",
     "fs-extra": "^9.0.1"
   }
 }
